# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/users": {
      "get": {
        "operationId": "AdminUsers_list",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "1-based page number.",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "Page size (max 100).",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search in email and full name (case-insensitive).",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ADMIN",
                "USER"
              ]
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "plan",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "FREE",
                "PREMIUM"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AdminUserDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMetaDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed to perform this action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List users (search by email/name, filter by role, status, plan)",
        "tags": [
          "Admin: Users"
        ]
      }
    },
    "/api/v1/admin/users/{id}": {
      "get": {
        "operationId": "AdminUsers_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed to perform this action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "User details with subscription, usage and activity counts",
        "tags": [
          "Admin: Users"
        ]
      },
      "patch": {
        "description": "Deactivation revokes all sessions immediately. Admins cannot modify themselves, and the last active admin cannot be demoted or deactivated.",
        "operationId": "AdminUsers_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed to perform this action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Change a user’s role or activate/deactivate the account",
        "tags": [
          "Admin: Users"
        ]
      },
      "delete": {
        "operationId": "AdminUsers_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted"
          },
          "400": {
            "description": "Validation failed or malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed to perform this action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Permanently delete a user and all of their data",
        "tags": [
          "Admin: Users"
        ]
      }
    },
    "/api/v1/admin/users/{id}/revoke-sessions": {
      "post": {
        "operationId": "AdminUsers_revokeSessions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed to perform this action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Sign a user out of every device",
        "tags": [
          "Admin: Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "EchoGPT API",
    "description": "\nREST API powering the **EchoGPT** multi-AI chat Chrome extension.\n\n### Authentication\n1. `POST /api/v1/auth/register` or `POST /api/v1/auth/login` returns an **access token** (short lived JWT) and a **refresh token** (opaque, long lived, rotated on every use).\n2. Send the access token as `Authorization: Bearer <token>`. Click **Authorize** above to use it here.\n3. When the access token expires call `POST /api/v1/auth/refresh`. Re-using an already rotated refresh token revokes the whole session.\n\n### Errors\nEvery error uses the same envelope: `{ statusCode, error, message, details?, path, timestamp, requestId }`.\nThe `x-request-id` response header echoes the correlation id for support/debugging.\n\n### Quotas\nAI endpoints (chat, AI-assisted search) consume the caller's daily plan quota. Exceeding it returns **429**.\n",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT access token"
      }
    },
    "schemas": {
...
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "3f6c1a2e-8a41-4a57-9d0f-1b8c0f2e6a11"
          },
          "email": {
            "type": "string",
            "example": "jane.doe@example.com"
          },
          "fullName": {
            "type": "string",
            "nullable": true,
            "example": "Jane Doe"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://cdn.example.com/avatar.png"
          },
          "role": {
            "enum": [
              "ADMIN",
              "USER"
            ],
            "type": "string",
            "example": "USER"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "emailVerified": {
            "type": "boolean",
            "example": false
          },
          "lastLoginAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "fullName",
          "avatarUrl",
          "role",
          "isActive",
          "emailVerified",
          "lastLoginAt",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "details": {
            "example": [
              "email must be an email",
              "password must be longer than or equal to 8 characters"
            ],
            "description": "Field level details, present for validation errors.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "path": {
            "type": "string",
            "example": "/api/v1/auth/register"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-09-29T10:15:30.000Z"
          },
          "requestId": {
            "type": "string",
            "example": "6f1c2c1e-6c47-4f0e-9a36-0c5f0f4b1a2d"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "path",
          "timestamp",
          "requestId"
        ]
      },
...
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "PlanResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 2
          },
          "code": {
            "enum": [
              "FREE",
              "PREMIUM"
            ],
            "type": "string",
            "example": "PREMIUM"
          },
          "name": {
            "type": "string",
            "example": "Premium"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "For power users."
          },
          "priceCents": {
            "type": "number",
            "description": "Monthly price in minor units (cents)",
            "example": 999
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "dailyRequestLimit": {
            "type": "number",
            "description": "AI requests allowed per UTC day",
            "example": 500
          },
          "features": {
            "example": [
              "500 AI requests per day"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "description",
          "priceCents",
          "currency",
          "dailyRequestLimit",
          "features"
        ]
      },
      "SubscriptionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "enum": [
              "ACTIVE",
              "CANCELED",
              "EXPIRED"
            ],
            "type": "string",
            "example": "ACTIVE"
          },
          "startedAt": {
            "format": "date-time",
            "type": "string"
          },
          "currentPeriodEnd": {
            "format": "date-time",
            "type": "string",
            "nullable": true,
            "description": "End of the paid period. Null for plans that never expire (Free)."
          },
          "canceledAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "plan": {
            "$ref": "#/components/schemas/PlanResponseDto"
          }
        },
        "required": [
          "id",
          "status",
          "startedAt",
          "currentPeriodEnd",
          "canceledAt",
          "plan"
        ]
      },
      "UsageResponseDto": {
        "type": "object",
        "properties": {
          "period": {
            "type": "string",
            "example": "DAILY",
            "description": "Quota window (UTC day)"
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "used": {
            "type": "number",
            "example": 3
          },
          "remaining": {
            "type": "number",
            "example": 17
          },
          "resetsAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the quota resets",
            "example": "2026-09-30T00:00:00.000Z"
          }
        },
        "required": [
          "period",
          "limit",
          "used",
          "remaining",
          "resetsAt"
        ]
      },
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 57
          },
          "totalPages": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ]
      },
...
      "AdminUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "3f6c1a2e-8a41-4a57-9d0f-1b8c0f2e6a11"
          },
          "email": {
            "type": "string",
            "example": "jane.doe@example.com"
          },
          "fullName": {
            "type": "string",
            "nullable": true,
            "example": "Jane Doe"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://cdn.example.com/avatar.png"
          },
          "role": {
            "enum": [
              "ADMIN",
              "USER"
            ],
            "type": "string",
            "example": "USER"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "emailVerified": {
            "type": "boolean",
            "example": false
          },
          "lastLoginAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "plan": {
            "nullable": true,
            "enum": [
              "FREE",
              "PREMIUM"
            ],
            "type": "string",
            "example": "FREE"
          }
        },
        "required": [
          "id",
          "email",
          "fullName",
          "avatarUrl",
          "role",
          "isActive",
          "emailVerified",
          "lastLoginAt",
          "createdAt",
          "updatedAt",
          "plan"
        ]
      },
      "AdminUserStatsDto": {
        "type": "object",
        "properties": {
          "conversations": {
            "type": "number",
            "example": 12
          },
          "messages": {
            "type": "number",
            "example": 148
          },
          "searches": {
            "type": "number",
            "example": 37
          },
          "activeSessions": {
            "type": "number",
            "example": 2
          }
        },
        "required": [
          "conversations",
          "messages",
          "searches",
          "activeSessions"
        ]
      },
      "AdminUserDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "3f6c1a2e-8a41-4a57-9d0f-1b8c0f2e6a11"
          },
          "email": {
            "type": "string",
            "example": "jane.doe@example.com"
          },
          "fullName": {
            "type": "string",
            "nullable": true,
            "example": "Jane Doe"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://cdn.example.com/avatar.png"
          },
          "role": {
            "enum": [
              "ADMIN",
              "USER"
            ],
            "type": "string",
            "example": "USER"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "emailVerified": {
            "type": "boolean",
            "example": false
          },
          "lastLoginAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "subscription": {
            "$ref": "#/components/schemas/SubscriptionResponseDto"
          },
          "usage": {
            "$ref": "#/components/schemas/UsageResponseDto"
          },
          "stats": {
            "$ref": "#/components/schemas/AdminUserStatsDto"
          }
        },
        "required": [
          "id",
          "email",
          "fullName",
          "avatarUrl",
          "role",
          "isActive",
          "emailVerified",
          "lastLoginAt",
          "createdAt",
          "updatedAt",
          "subscription",
          "usage",
          "stats"
        ]
      },
      "AdminUpdateUserDto": {
        "type": "object",
        "properties": {
          "role": {
            "enum": [
              "ADMIN",
              "USER"
            ],
            "type": "string",
            "example": "ADMIN"
          },
          "isActive": {
            "type": "boolean",
            "example": false,
            "description": "Deactivating a user immediately revokes all of their sessions."
          }
        }
      },
...
    }
  }
}